Create Simple Integer Calculator.c
This is my first mini-mini project in my journey to becoming a Software Developer! I am currently using and learning the C programming language along with basic computer science concepts.

This simple integer calculator was made so that the user can continue to enter math problems and get the result as many times as they like. I did so by including a switch case within a while loop.